+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 1.2.2 - WIP
+## 1.2.2 - 2026-02-21
 - Adding `method` parameter to `quantiles()` supporting `'exclusive'` (default) and `'inclusive'` interpolation methods
 - Adding `medianGrouped()` method for estimating the median of grouped/binned continuous data using interpolation
 - Adding Spearman rank correlation via `method` parameter in `correlation()` (`method='ranked'`)